LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/claim-with-redirect.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Cannot update a null/nonexistent service worker registration
  

FAIL Claim works after redirection to another origin assert_equals: expected (string) "updated" but got (undefined) undefined
